scraping/requirements.txt
python-dwca-reader
requests
pandas
jupyter
ipykernel